Accessibility fixes
Setting ariaLabel and ariaRole properties with TypeScript seem not to
change the HTML aria-label and aria-role properties. In addition to
other changes, such attributes are now set using the setAttribute
method.

Further, a textarea for input has been added -- when not in a textarea,
keyboard input is (in Firefox) interpreted as commands for navigating
the page.