fix(toolbar): Select added digits if the document size input changes from zero while editing
Previously, trying to change the value of a dimension input that ends in
0s (e.g. from 100 to 200) by deleting and replacing the first digit
caused the input's value to be reset to 100 (as 0 is an invalid
dimension).

This commit works around the issue by selecting the changed digit(s) at
the beginning of the input.

This should fix #58 (in many cases).