feat(editor): Pasting: Interpret SVG text/plain data as SVG
In many places, the MIME types available for copy/pate are limited. For example,
in a WebView running in an Android React Native app, an app might work
around the lack of a working WebView clipboard API by using
`@react-native-clipboard/clipboard`, which currently doesn't seem to
support copying HTML or SVG. This commit will allow js-draw data to be
copy/pasted as text/plain.