fix: Chrome+Linux: Fix touchscreen zooming in Chromium 138
This change works around a Chromium bug
(https://issues.chromium.org/u/1/issues/428153664) in which pointer
events from a touchscreen were given an empty string for pointerType. In
this case, js-draw marked these pointers as of type "Other" and did not
allow multi-pointer zoom.

This change allows pointers of type Other to pinch-zoom.